Separated file validation from fixtures to avoid redundant test runs
