+++ gents/tests/test_cases.py
@@ -59,7 +59,6 @@ def generate_history_file(path, time_val, time_bounds_val, num_vars=SIMPLE_NUM_V
     
     ds.setncatts({
         "source": "GenTS testing suite",
-        "time_created": get_time_stamp(),
         "description": "Synthetic data used for testing with the GenTS package.",
         "frequency": "month",
     })
@@ -108,22 +107,3 @@ def structured_case(tmp_path_factory):
                 path = f"{dir_path}/testing.hf.{str(file_index).zfill(5)}.nc" 
                 generate_history_file(path, (file_index+1)*30, [file_index*30, (file_index+1)*30], num_vars=SCRAMBLED_NUM_VARS)
     return head_hf_dir, head_ts_dir
-
-
-def test_simple_case(simple_case):
-    input_head_dir, output_head_dir = simple_case
-    assert len(listdir(input_head_dir)) == SIMPLE_NUM_TEST_HIST_FILES
-
-
-def test_scrambled_case(scrambled_case):
-    input_head_dir, output_head_dir = scrambled_case
-    assert len(listdir(input_head_dir)) == SCRAMBLED_NUM_TEST_HIST_FILES
-
-
-def test_structured_case(structured_case):
-    input_head_dir, output_head_dir = structured_case
-    assert len(listdir(input_head_dir)) == STRUCTURED_NUM_DIRS
-    for top_dir in listdir(input_head_dir):
-        assert len(listdir(f"{input_head_dir}/{top_dir}")) == STRUCTURED_NUM_SUBDIRS
-        for sub_dir in listdir(f"{input_head_dir}/{top_dir}/"):
-            assert len(listdir(f"{input_head_dir}/{top_dir}/{sub_dir}")) == STRUCTURED_NUM_TEST_HIST_FILES
